Reworked the logic of the bot

+++ GusMelfordBot.Infrastructure/Models/MetaContent.cs
@@ -5,6 +5,6 @@ public class MetaContent : AuditableEntity
     public int? Height { get; set; }
     public int? Width { get; set; }
     public int? Duration { get; set; }
-    public long? TelegramMessageId { get; set; }
+    public int? TelegramMessageId { get; set; }
     public bool IsSaved { get; set; }
 }
